default project template to 'blank'
UNTESTED! (but seems easy enough)

Why is the default layout "tabs" and not "blank"?

It's great to have some more advanced starting themes, but why is the default/fallback "tabs"?

It is important to have some kind of display to don't seem broken, but "blank" has this. 

...I am new to ionic and stumbled across this question while doing the 'getting started' guide where I had to delete the tabs thing to have a clean start. 

I have checked the code back to the introduction of the app themes and didn't find a reason for "tabs" there.