fix: prefer locally installed executables
In particular, for the case where one has a locally installed `cordova`
package, an executable is placed in `node_modules/.bin`, but that PATH
is not recognized by the Ionic CLI--only a globally installed Cordova.
Now the Ionic CLI will detect and prefer locally installed executables.

This mimics npm behavior (think of npm scripts), and should've been the
default behavior of the Ionic CLI. Therefore, it is marked as a fix, not
a feat.

fixes https://github.com/ionic-team/ionic-cli/issues/1132