re-architect tasks
Keeping BottomBar open seems to be a consistent issue on Windows.
BottomBar creates a readline interface, so SIGINT has to go through it,
and the signal doesn't seem to bubble up properly, which means the Ionic
CLI cannot properly clean up resources during a Ctrl-C, leading to serve
issues where the underlying CLI is not terminated.