feat(capacitor): add run command
`ionic capacitor run` is similar to `ionic cordova run` in that it
invokes an Ionic build or serve (depending on the `--livereload` option)
and copies assets and configuration to native projects. It differs in
that it does not programmatically run the native project in a device or
emulator/simulator. Capacitor projects must still use the Native IDE
(Xcode or Android Studio) to run apps natively (for now :wink:).

Resolves https://github.com/ionic-team/ionic-cli/issues/3130