feat(login): SSO authentication flow (#3741)
This is a feature for Ionic Pro users whose org has SSO enabled. At this point, it is mostly a proof of concept. It is hidden behind a flag and should not affect standard email/password authentication.

One known issue is that tokens (at the time of writing) expire in 1 day. API access was originally implemented with the idea that we had a long-lived token. We will need to utilize the refresh token to generate a new access token as needed.