chore(serve): rename `--address` option to `--host`
The term "address" is ambiguous and could mean host, host & port, or
even a full URL. It could also be interpreted as IP address, which is
not necessarily the case with `ionic serve`.

Additionally, this aligns the Ionic CLI with popular tools such as
Webpack Dev Server, Angular CLI, React Scripts, etc. which all call it
host.

A deprecation message will be shown when `--address` is used.