Update example commands to use "Signing Certificate"
References to the term "security profile" in Appflow and the associated documentation are changing to "signing certificate". This change update the references to this in the example commands without changing the name of the actual flag in hopes to make this language change more intuitive for the user following the documentation.