fix(#3965): give --lab access to serve externalAddress
This should fix a bug where lab is trying to always grab content
from localhost even if the content is hosted on a different address.

This should help with accessing lab on an external network in a browser
when say for extample the ionic-cli is running on ubuntu server, where a
browser is not available.

The reason we will use `serveDetails.externalAddress` instead of
`labDetails.host` is because `serve` is where we will be getting our content
from and by default is hosted on a seperate port from lab.

This change should not break anything since `serveDetails.externalAddress`
will default to the initial `localhost` if no --host argument is passed.